‘简易的登录和valve。’

+++ dream-dal/src/main/java/com/sven/dream/base/entity/DreamDo.java
@@ -4,7 +4,7 @@ import java.util.Date;
 
 public class DreamDo {
 	 
-    private Integer id;
+    private Long id;
  
     private Date gmtCreate;
  
@@ -15,12 +15,12 @@ public class DreamDo {
     private String modifier;
  
     private String isDeleted;
-
-	public Integer getId() {
+    
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
